Create cpugov
A few days ago I was considering the lack of CPU throttling utilities for Linux. Since I have an old power hogging FX-8350, I wrote this script. It is not exactly a throttling tool, but it works well for what I need out of it. This is getting to the limit of my scripting skills. I am sure it can be improved and expanded on, so here you go.